updates badges labels, adds to customer io payload, adds as an array to the model and adds as a fillable field
